Flask/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
marshmallow==3.20.1
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
bcrypt==4.0.1
email-validator==2.0.0 